FEATURE: update imports accordingly renames
